@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,9 +27,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  agora_rtc_engine:
-  permission_handler:
+  cupertino_icons: ^1.0.3
+  permission_handler: ^8.1.4+2
+  agora_rtc_engine: ^4.0.6
 
 dev_dependencies:
   flutter_test:
